display interactivity improvements, accessibility and compat with scrolling Apr 28
